TS migration of remaining JS components
